Fixes for WLM logs

+++ include/Opensea_Parser_Helper.h
@@ -323,7 +323,7 @@ namespace opensea_parser {
         }
         else
         {
-            if ((int64_t)value < INT32_MAX && (int64_t)value > INT32_MIN)
+            if ((int32_t)value < INT32_MAX && (int32_t)value > INT32_MIN)
             {
                 json_push_back(nowNode, json_new_i((char *)myStr.c_str(), static_cast<int32_t>(M_DoubleWord0(value))));
             }
